Rename previous 'users-and-permissions' systems directory to 'users'
